Skyline: Improved handling of negative ion mode chromatograms for small molecule work.  No longer tries to match negative ion mode chromatogram or scan data with positive transitions, and vice versa. This change introduces the SignedMz struct - a means of distinguishing m/z values for negative ion modes from those for positive.  Some notes on that: It's tempting to just carry around negative values to indicate negative ion mode m/z values, but this is complicated by the fact that m/z is inherently "positive", in the sense of sorting etc, and the code would have to be peppered with special case code to deal with this.  Instead we have this new object that's specifically for m/z data, which carries around the m/z value and a sign flag, and when we sort it's by sign then by m/z ascending.  There's a lot of operator magic in SignedMz to minimize code changes where we used to implement mz values as simple doubles.
Former-commit-id: b6a6df49f4399aaece41048a8d468aacb9594b00

+++ pwiz_tools/Skyline/Model/PeakMatcher.cs
@@ -87,7 +87,7 @@ namespace pwiz.Skyline.Model
             var abundances = new IonAbundances();
             foreach (var nodeTran in nodeTranGroup.Transitions)
             {
-                var chromInfoCached = chromGroupInfo.GetTransitionInfo((float) nodeTran.Mz, mzMatchTolerance);
+                var chromInfoCached = chromGroupInfo.GetTransitionInfo(nodeTran.Mz, mzMatchTolerance);
                 if (chromInfoCached == null)
                     continue;
 
@@ -295,7 +295,7 @@ namespace pwiz.Skyline.Model
             var largestPeak = ChromPeak.EMPTY;
             foreach (var peak in
                      from transitionDocNode in nodeTranGroup.Transitions
-                     select chromGroupInfo.GetTransitionInfo((float) transitionDocNode.Mz, mzMatchTolerance)
+                     select chromGroupInfo.GetTransitionInfo(transitionDocNode.Mz, mzMatchTolerance)
                      into chromInfo where chromInfo != null
                      select chromInfo.GetPeak(peakIndex))
             {
@@ -504,7 +504,7 @@ namespace pwiz.Skyline.Model
             {
                 foreach (var nodeTran in nodeTranGroup.Transitions)
                 {
-                    var chromInfoCached = chromGroupInfo.GetTransitionInfo((float) nodeTran.Mz, mzMatchTolerance);
+                    var chromInfoCached = chromGroupInfo.GetTransitionInfo(nodeTran.Mz, mzMatchTolerance);
                     if (chromInfoCached == null)
                         continue;
